expand man page
In the DESCRIPTION section, the interaction between gnupg-pkcs11-scd and
GPG has been expanded.

In the GNUPG INTEGRATION section, an example of adding keys as subkeys
on an existing identity is described.

Modify final integration step regarding emulate-openpgp. I don't have
access to a card that requires disabling `emulate-openpgp`, and my card
requires emulation in order to work with GPG. There may be an
implementation divergence here.