gnupg-pkcs11-scd: add PKAUTH command support
This commit adds PKAUTH command support to gnupg-pkcs11-scd, so
gpg-agent with "enable-ssh-support" setting defined can act as a SSH
Agent when PKCS11 is used as a GPG backend.

Auth operation is almost the same as sign operation, except it looks
like SSH always sends data with hash algorithm signature appended at the
beginning, but the data is of different size than the signature
detection code expects, so it always fallback to default behavior, which
is to append SHA1 signature.

As having 2 different signature prefixes is incorrect, we need to
use the different default value from the sign action, which is to not
append anything in case of auth operation.

Closes #27

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>